feat: 支持可微分 VAE 编码并修复精度问题
- 在 pipeline 中实现 _encode_vae_image_differentiable 以支持梯度回传
- LocalGuidance 切换为可微分编码模式，并在 loss 计算时强制转为 float32 以避免 bf16 backward 错误
- 更新训练脚本使用的 GPU 和任务名称

+++ edit4shape/guidance/flowedit/state_tracker.py
@@ -75,7 +75,8 @@ class FlowEditStateTracker:
         """
         # rendered: [B, seq_len, C] packed
         # self.latents[k]: [B, seq_len, C] packed
-        return torch.stack([F.mse_loss(rendered, lat) for lat in self.latents])  # [K]
+        # 转 float32 避免 bfloat16 backward 报错
+        return torch.stack([F.mse_loss(rendered.float(), lat.float()) for lat in self.latents])  # [K]
     
     def loss_final(self, rendered: torch.Tensor) -> torch.Tensor:
         """
@@ -84,7 +85,7 @@ class FlowEditStateTracker:
         Args:
             rendered: 渲染图的 packed latent [B, seq, C]
         """
-        return F.mse_loss(rendered, self.final)
+        return F.mse_loss(rendered.float(), self.final.float())
     
     def loss_mean(self, rendered: torch.Tensor) -> torch.Tensor:
         """
